Adjust game link retrieval to use clipboard
Updated the game creation process to accommodate changes on the website Spotle.io, which now passes the game link to the user's clipboard instead of through browser logs. Deprecated the old log-based code extraction method and introduced a new function to extract the link directly from the clipboard using regex, simplifying the process and improving reliability.